fix: las ventas con centavos ya cuadran con sus lineas (v1.13.0)
producto_venta.precio_unitario era INTEGER mientras ventas.total es
NUMERIC(10,2). Tres unidades a C$10.10 se facturaban como C$30.30 pero las
lineas sumaban C$30.00, y como el saldo sale de las lineas el cobro correcto
se rechazaba por "excede el saldo pendiente".

De los tres caminos que registran una venta solo el POS redondeaba a
proposito; la entrega de orden de taller y la conversion de cotizacion
perdian los centavos en silencio al pasarle un Decimal a la columna entera.

Ademas, en convertir_venta el total se calculaba antes de redondear los
precios de linea, asi que podian discrepar por centavos (C$100 entre 3 son
C$33,3333...). Ahora se redondea primero y el total se suma de ahi.

Migracion 0025 (ALTER COLUMN a NUMERIC(10,2)), validada en base desechable
aplicando/revirtiendo/reaplicando y ya aplicada en produccion con respaldo
previo verificado. No transforma datos.

El centavo que perdieron las ventas anteriores no quedo guardado en ninguna
parte: se agrega `manage.py verificar_totales_venta` para listarlas en vez
de inventar el dato.

347 tests, todos pasando.

Generated-By: PostHog Code
Task-Id: 90d658e3-5d6b-49bb-90c7-ce9081565d93

+++ backend/SQL_FILES/000_base_schema_snapshot.sql
@@ -2,7 +2,8 @@
 -- SNAPSHOT DE ESQUEMA BASE (solo estructura, sin datos)
 -- Generado con: pg_dump --schema-only --no-owner --no-privileges
 --   --exclude-table=<tablas administradas por Django, ver mas abajo>
--- Fecha: 2026-07-24 (regenerado tras mitigar R03: unificacion de esquema)
+-- Fecha: 2026-08-06 (producto_venta.precio_unitario pasa a numeric, mig. 0025)
+-- Anterior: 2026-07-24 (regenerado tras mitigar R03: unificacion de esquema)
 --
 -- Este archivo NO se ejecuta contra la base de datos de produccion/dev
 -- (las tablas ya existen ahi). Su proposito es:
@@ -1120,7 +1121,7 @@ CREATE TABLE public.producto_venta (
     id_venta integer NOT NULL,
     id_producto integer NOT NULL,
     cantidad integer NOT NULL,
-    precio_unitario integer NOT NULL
+    precio_unitario numeric(10,2) NOT NULL
 );
